ATT003

+++ resources/views/emprestimo/show.blade.php
@@ -37,14 +37,14 @@
             <p class="text">obs: {{$emprestimo->obs}}</p>
         </div>
         <div class="card-footer">
-            @auth
+            @if ((Auth::check()) && (Auth::user()->isAdmin()))
                 {{Form::open(['route' => ['emprestimos.destroy',$emprestimo->id],'method' => 'DELETE'])}}
                 {{Form::submit('Excluir',['class'=>'btn btn-danger','onclick'=>'return confirm("Confirma exclusão?")'])}}
-            @endauth
+            @endif
             <a href="{{url('emprestimos/')}}" class="btn btn-warning">Voltar</a>
-            @auth
+            @if ((Auth::check()) && (Auth::user()->isAdmin()))
                 {{Form::close()}}
-            @endauth
+            @endif
           
         </div>
     </div>
